requirements.txt
django==1.6
requests==2.0.1
django-pipeline==1.3.15
django-extensions==1.2.5
Pillow==2.2.1
easy-thumbnails==1.4
chaussette==1.0
circus==0.10.0
psycopg2==2.5.1
slimit==0.8.1
isodate==0.4.9
djangorestframework==2.3.9
